Review comments plus enhancements to StringFWTest.

+++ src/test/java/org/reaktivity/nukleus/maven/plugin/internal/ast/parse/AstParserTest.java
@@ -207,7 +207,7 @@ public class AstParserTest
                 .build();
 
         assertEquals(expected, actual);
-        assertEquals(-12, actual.defaultValue().intValue());
+        assertEquals(-12, actual.defaultValue());
     }
 
     @Test
